add class EventCodeClient with function loadEventMaps, update ConditionCode data class, rename java folders to kotlin, add unit tests for RedisUtility and EventCodeClient, add resource directory event_codes and csv files for arbo-lyme and hep
