+++ mrr/src/test/kotlin/gov/cdc/dex/mrr/EventCodeClientTest.kt
@@ -1,0 +1,14 @@
+package gov.cdc.dex.mrr
+
+import org.junit.jupiter.api.Test
+
+import org.junit.jupiter.api.Assertions.*
+
+internal class EventCodeClientTest {
+
+    @Test
+    fun loadEventMapsTest() {
+        var event_codes : EventCodeClient = EventCodeClient()
+        event_codes.loadEventMaps()
+    }
+}
